requirements.txt
botoy
httpx
httpx_socks
pyzipper
pydantic
Pillow
tenacity
